Move graph functionality to graph component
